Moved `Ammo` class into a separated file
